requirements.txt
# Django framework
Django>=4.2

# Vector database
chromadb

# Text embedding and splitting
sentence-transformers
langchain

# Numerical computing
numpy

# Hugging Face LLM
transformers
torch   # Required backend for Hugging Face models

#PDF and Word doc parsing for uploads
pdfplumber
python-docx

#For async processing or web requests if needed
requests

djangorestframework>=3.15

# for pdf/text parsing
PyPDF2

# for cleaning text and removing non-ASCII characters and being able to read charts
pdfplumber
# to run all do this  
# pip install -r requirements.txt